refactor: extract class which calculates max value of pages
