+++ src/main/java/onboarding/Problem1.java
@@ -1,5 +1,6 @@
 package onboarding;
 
+import onboarding.problem1.PageCalculator;
 import onboarding.problem1.PageChecker;
 
 import java.util.List;
@@ -17,6 +18,7 @@ class Problem1 {
     private final int SIZE;
 
     private final PageChecker checker;
+    private final PageCalculator calculator;
 
     /**
      * Constructor with pages of person A and B
@@ -29,6 +31,7 @@ class Problem1 {
         this.pagesB = pagesB;
         this.SIZE = size;
         checker = new PageChecker(size);
+        calculator = new PageCalculator();
     }
 
     /**
@@ -38,7 +41,7 @@ class Problem1 {
      */
     public int winner() throws WrongInputException {
         if (checker.isAvailableInput(pagesA) && checker.isAvailableInput(pagesB))
-            return resultOfGame(valueOfPages(pagesA.get(0), pagesA.get(1)), valueOfPages(pagesB.get(0), pagesB.get(1)));
+            return resultOfGame(calculator.valueOfPages(pagesA.get(0), pagesA.get(1)), calculator.valueOfPages(pagesB.get(0), pagesB.get(1)));
         else
             throw new WrongInputException("Wrong input");
     }
@@ -52,54 +55,7 @@ class Problem1 {
         }
     }
 
-    /**
-     * Max value that can be calculated with page numbers
-     * @param left left page number
-     * @param right right page number
-     * @return max value
-     */
-    private int valueOfPages(int left, int right) {
-        return Math.max(valueOfPage(left), valueOfPage(right));
-    }
 
-    /**
-     * Max value that can be calculated with page number
-     * @param page page number
-     * @return max value
-     */
-    private int valueOfPage(int page) {
-        return Math.max(plusEach(page), multiplyEach(page));
-    }
-
-    /**
-     * Summation of each number in page number
-     * @param page page number
-     * @return summation
-     */
-    private int plusEach(int page) {
-        int sum = 0;
-        while (page != 0) {
-            sum += page % 10;
-            page /= 10;
-        }
-        return sum;
-    }
-
-    /**
-     * Multiplied value of each number in page number
-     * @param page page number
-     * @return multiplied value
-     */
-    private int multiplyEach(int page) {
-        int mul = 1;
-        while (page != 0) {
-            mul *= page % 10;
-            page /= 10;
-            if (mul == 0)
-                break;
-        }
-        return mul;
-    }
 
     /**
      * Get result of game
